add: new player model with fixes
