+++ Assets/Animations/Box/FoodBoxAnim.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7ee909a28fc04541b5963b8cfdff25b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
