+++ Assets/Scripts/GameController.cs
@@ -11,6 +11,6 @@ public class GameController : MonoBehaviour
 
     void Update()
     {
-
+        
     }
 }
