--- Assets/Scripts/PlayerMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d1e1558259d268a47aa9436cae2c8430
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
