+++ ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Food
   - NormalBlocks
   - InteractableBlocks
+  - FoodBox
   layers:
   - Default
   - TransparentFX
